pyproject.toml
[project]
name = "tradestation-bot"
version = "0.1.0"
description = ""
requires-python = ">=3.13"
dependencies = [
    "requests>=2.33.1,<3.0.0",
    "python-dotenv>=1.2.2,<2.0.0"
]

[tool.poetry]
packages = [{ include = "app" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"